// resources/js/types/ProveedoresFormValues.ts
export type ProveedoresFormValues = {
    tipo: string
    nombre: string
    apellido_paterno: string
    apellido_materno: string
    razon_social: string
    nombre_comercial: string
    numero_exterior: string
    numero_interior: string
    pagina_web: string
    descripcion: string
    estatus: boolean
    rfc: string
    calle: string
    colonia: string
    municipio_id: string
    codigo_postal: string
    telefono: string
    correo: string
}